+++ graal/com.oracle.graal.truffle/src/com/oracle/graal/truffle/DefaultTruffleSplittingStrategyNew.java
@@ -51,7 +51,7 @@ public class DefaultTruffleSplittingStrategyNew implements TruffleSplittingStrat
         if (!TruffleCompilerOptions.TruffleSplitting.getValue()) {
             return false;
         }
-        if (!call.isSplittable()) {
+        if (!call.isCallTargetCloningAllowed()) {
             return false;
         }
         if (TruffleCompilerOptions.TruffleSplittingAggressive.getValue()) {
@@ -80,8 +80,8 @@ public class DefaultTruffleSplittingStrategyNew implements TruffleSplittingStrat
         OptimizedCallTarget currentTarget = call.getCurrentCallTarget();
 
         if (splittingForced) {
-            if (!call.isSplit()) {
-                call.installSplitCallTarget(currentTarget.split());
+            if (!call.isCallTargetCloned()) {
+                call.installSplitCallTarget(currentTarget.cloneUninitialized());
             }
             return;
         }
@@ -123,7 +123,7 @@ public class DefaultTruffleSplittingStrategyNew implements TruffleSplittingStrat
                 // we are the only caller + the profile is not polluted by other call sites
                 // -> reuse the currentTarget but update the profile if necessary
                 newTarget = currentTarget;
-                if (currentTarget.getSplitSource() != null) {
+                if (currentTarget.getSourceCallTarget() != null) {
                     profiles.remove(oldProfile);
                     profiles.put(newProfile, newTarget);
                 }
@@ -131,7 +131,7 @@ public class DefaultTruffleSplittingStrategyNew implements TruffleSplittingStrat
                 newTarget = profiles.get(newProfile);
                 if (newTarget == null) {
                     // in case no compatible target was found we need to split
-                    newTarget = target.split();
+                    newTarget = target.cloneUninitialized();
                     profiles.put(newProfile, newTarget);
                     split = true;
                 }
@@ -141,7 +141,7 @@ public class DefaultTruffleSplittingStrategyNew implements TruffleSplittingStrat
         call.installSplitCallTarget(newTarget);
 
         if (split && TruffleCompilerOptions.TraceTruffleSplitting.getValue()) {
-            traceSplit(currentTarget.getSplitSource() != null ? oldProfile : currentTarget.getArgumentStamp(), newProfile);
+            traceSplit(currentTarget.getSourceCallTarget() != null ? oldProfile : currentTarget.getArgumentStamp(), newProfile);
         }
 
         cleanup(currentTarget);
@@ -151,7 +151,7 @@ public class DefaultTruffleSplittingStrategyNew implements TruffleSplittingStrat
     private void traceSplit(TruffleStamp oldStamp, TruffleStamp newStamp) {
         OptimizedCallTarget callTarget = call.getCallTarget();
         Map<TruffleStamp, OptimizedCallTarget> splitTargets = callTarget.getSplitVersions();
-        String label = String.format("split %3s-%-4s-%-4s ", splitChangeCount++, call.getCurrentCallTarget().getSplitIndex(), call.getCallCount());
+        String label = String.format("split %3s-%-4s-%-4s ", splitChangeCount++, call.getCurrentCallTarget().getCloneIndex(), call.getCallCount());
         OptimizedCallTargetLog.log(0, label, callTarget.toString(), callTarget.getDebugProperties());
         logProfile(callTarget.getArgumentStamp(), callTarget, oldStamp, newStamp);
         for (TruffleStamp profile : splitTargets.keySet()) {
@@ -160,7 +160,7 @@ public class DefaultTruffleSplittingStrategyNew implements TruffleSplittingStrat
     }
 
     private static void logProfile(TruffleStamp stamp, OptimizedCallTarget target, TruffleStamp oldStamp, TruffleStamp newStamp) {
-        String id = String.format("@%8h %s", target.hashCode(), target.getSplitSource() == null ? "orig." : "split");
+        String id = String.format("@%8h %s", target.hashCode(), target.getSourceCallTarget() == null ? "orig." : "split");
         String plusMinus = stamp.equals(newStamp) ? "+ " : (stamp.equals(oldStamp) ? "- " : "");
         System.out.printf("%16s%-20sCallers: %3d, Nodes:%10s %s%n", plusMinus, id, target.getKnownCallSiteCount(), //
                         String.format("%d (%d/%d)", count(target, NodeCost.MONOMORPHIC), count(target, NodeCost.POLYMORPHIC), count(target, NodeCost.MEGAMORPHIC)),//
@@ -176,8 +176,8 @@ public class DefaultTruffleSplittingStrategyNew implements TruffleSplittingStrat
     }
 
     private static void cleanup(OptimizedCallTarget currentTarget) {
-        if (currentTarget.getKnownCallSiteCount() == 0 && currentTarget.getSplitSource() != null) {
-            OptimizedCallTarget removed = currentTarget.getSplitSource().getSplitVersions().remove(currentTarget.getArgumentStamp());
+        if (currentTarget.getKnownCallSiteCount() == 0 && currentTarget.getSourceCallTarget() != null) {
+            OptimizedCallTarget removed = currentTarget.getSourceCallTarget().getSplitVersions().remove(currentTarget.getArgumentStamp());
             if (removed != null) {
                 disposeTarget(removed);
             }